addressed changes + refactored page.tsx into separate pages to reduce bloat
